Return format raw as string
If the template engine gets a []byte object as value, it will format it when returning it later as a string, making it hard to convert it back to []byte correctly. Therefore, before the template engine butchers our raw value, we encode it as a string before so that it can be converted back later on correctly.