AB#361 Change manifest placeholders to Go templates (#70)
* Use templates instead of placeholder strings in manifest

* Wrap secrets to have "Marblerun" as prefix

* Remove placeholder variables

* Access a symmetric key through ".Key" in template

* Rename & Refactor

* Update example manifest in BUILD.md

* Return format raw as string

If the template engine gets a []byte object as value, it will format it when returning it later as a string, making it hard to convert it back to []byte correctly. Therefore, before the template engine butchers our raw value, we encode it as a string before so that it can be converted back later on correctly.